backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27

SQLAlchemy>=2.0
psycopg2-binary>=2.9

pandas>=2.1
python-dotenv>=1.0

openai>=1.0

# Memory (Phase 4)
redis>=5.0

# Vector DB (Phase 2)
# NOTE: Chroma currently depends on onnxruntime, which does not ship wheels for Python 3.14.
# FAISS has Windows wheels for Python 3.14, so we use it for local similarity search.
faiss-cpu>=1.13.2


